backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
yfinance==0.2.32
python-multipart==0.0.6
httpx==0.25.2
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
apscheduler==3.10.4
beautifulsoup4==4.12.2
praw==7.7.1
tweepy==4.14.0
